Add check file validation

+++ srcs/requirements/wordpress/tools/setup.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-echo -e "Download wordpress with wp-cli"
-wp core download --path=/var/www/wordpress
 
 echo -e "Make www user and give permissions"
 adduser -D www
@@ -19,14 +17,20 @@ echo "env[MARIADB_USER] = \$MARIADB_USER" >> /etc/php7/php-fpm.d/www.conf
 echo "env[MARIADB_PWD] = \$MARIADB_PWD" >> /etc/php7/php-fpm.d/www.conf
 echo "env[MARIADB_DB] = \$MARIADB_DB" >> /etc/php7/php-fpm.d/www.conf
 
-echo -e "Config wp-config.php"
-cp /tmp/wp-config.php /var/www/wordpress/wp-config.php
+if ! [[ -f /var/www/wordpress/wp-config.php ]];
+then
+    echo -e "Download wordpress with wp-cli"
+    wp core download --path=/var/www/wordpress
 
-echo -e "Install wordpress"
-wp core install --url="$WP_URL" --title="$WP_TITLE" --admin_user="$WP_ADMIN_USER" --admin_password="$WP_ADMIN_PWD" --admin_email="$WP_ADMIN_EMAIL" --skip-email --path=/var/www/wordpress
+    echo -e "Config wp-config.php"
+    cp /tmp/wp-config.php /var/www/wordpress/wp-config.php
 
-echo -e "Add default user"
-wp user create hyun hyun@example.com --user_pass=password --path=/var/www/wordpress
+    echo -e "Install wordpress"
+    wp core install --url="$WP_URL" --title="$WP_TITLE" --admin_user="$WP_ADMIN_USER" --admin_password="$WP_ADMIN_PWD" --admin_email="$WP_ADMIN_EMAIL" --skip-email --path=/var/www/wordpress
+
+    echo -e "Add default user"
+    wp user create ${WP_DEFAULT_USER} ${WP_DEFAULT_EMAIL} --user_pass=${WP_DEFAULT_PWD} --path=/var/www/wordpress
+fi
 
 echo -e "Start fpm server"
 php-fpm7 --nodaemonize
